Move match IR into its own file
Keep the statement compilation hook in ir.cpp, but move the native match operator, pattern lowering, validation, and inspection plugin setup into a dedicated match IR translation unit.
